requirements.txt
google-cloud-language==2.3.0
openai==0.27.0
transformers==4.13.0
